+++ compiler/GHC/Tc/Gen/Splice.hs
@@ -26,7 +26,7 @@ module GHC.Tc.Gen.Splice(
      runMetaE, runMetaP, runMetaT, runMetaD, runQuasi,
      tcTopSpliceExpr, lookupThName_maybe,
      defaultRunMeta, runMeta', runRemoteModFinalizers,
-     finishTH, runTopSplice
+     finishTH, runTopSplice, reifyName
       ) where
 
 import GHC.Prelude
@@ -2134,15 +2134,6 @@ reifyTyCon :: TyCon -> TcM TH.Info
 reifyTyCon tc
   | Just cls <- tyConClass_maybe tc
   = reifyClass cls
-
-{-  Seems to be just a short cut for the next equation -- omit
-  | tc `hasKey` fUNTyConKey -- I'm not quite sure what is happening here
-  = return (TH.PrimTyConI (reifyName tc) 2 False)
--}
-
-  | isPrimTyCon tc
-  = return (TH.PrimTyConI (reifyName tc) (length (tyConVisibleTyVars tc))
-                          (isUnliftedTypeKind (tyConResKind tc)))
 
   | isTypeFamilyTyCon tc
   = do { let tvs      = tyConTyVars tc