+++ libraries/ghc-boot-th/GHC/Internal/TH/Ppr.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE Trustworthy #-}
 {-# LANGUAGE LambdaCase #-}
+{-# OPTIONS_GHC -Wno-deprecations #-}
 -- | contains a prettyprinter for the
 -- Template Haskell datatypes